Rename to slot-and-fill
